+++ TA-jira-service-desk-simple-addon/bin/TA_jira_service_desk_simple_addon_rh_settings.py
@@ -130,6 +130,13 @@ fields_additional_parameters = [
         validator=None
     ),
     field.RestField(
+        'jira_ssl_certificate_path',
+        required=False,
+        encrypted=False,
+        default=0,
+        validator=None
+    ),
+    field.RestField(
         'jira_passthrough_mode',
         required=False,
         encrypted=False,