+++ TA-jira-service-desk-simple-addon/bin/jirafill.py
@@ -21,6 +21,7 @@ import ta_jira_service_desk_simple_addon_declare
 from splunklib.searchcommands import dispatch, GeneratingCommand, Configuration, Option, validators
 
 import sys
+import os
 import splunk
 import time
 import requests
@@ -41,11 +42,11 @@ class GenerateTextCommand(GeneratingCommand):
         else:
             return '%s/rest/api/latest/%s' % (url, endpoint)
 
-    def get_jira_info(self, username, password, url, proxy_dict , endpoint):
+    def get_jira_info(self, username, password, url, ssl_verify, proxy_dict , endpoint):
         response = requests.get(
             url=self.jira_url(url, endpoint),
             auth=(username, password),
-            verify=False,
+            verify=ssl_verify,
             proxies=proxy_dict
         )
         return response.json()
@@ -56,6 +57,8 @@ class GenerateTextCommand(GeneratingCommand):
         confs = self.service.confs[str(conf_file)]
         proxy_url = None
         proxy_dict = None
+        ssl_verify = False
+        ssl_cert_path = None
         for stanza in confs:
             if stanza.name == "additional_parameters":
                 for key, value in stanza.content.items():
@@ -63,6 +66,10 @@ class GenerateTextCommand(GeneratingCommand):
                         username = value
                     if key == "jira_url":
                         url = value
+                    if key == "jira_ssl_certificate_validation":
+                        jira_ssl_certificate_validation = value
+                    if key == "jira_ssl_certificate_path":
+                        ssl_cert_path = value
             if stanza.name == "proxy":
                 for key, value in stanza.content.items():
                     if key == "proxy_enabled":
@@ -80,7 +87,13 @@ class GenerateTextCommand(GeneratingCommand):
               "http" : proxy_url + ":" + proxy_port,
               "https" : proxy_url + ":" + proxy_port
               }
-
+        if jira_ssl_certificate_validation:
+            if jira_ssl_certificate_validation == '0':
+                ssl_verify = False
+            elif jira_ssl_certificate_validation == '1' and ssl_cert_path and os.path.isfile(ssl_cert_path):
+                ssl_verify = str(ssl_cert_path)
+            elif jira_ssl_certificate_validation == '1':
+                ssl_verify = True
 
         for credential in storage_passwords:
             if credential.content.get('username') == "additional_parameters``splunk_cred_sep``1" and credential.content.get('clear_password').find('jira_password') > 0:
@@ -88,22 +101,22 @@ class GenerateTextCommand(GeneratingCommand):
                 break
 
         if self.opt == 1:
-            for project in self.get_jira_info(username, password, url, proxy_dict ,'project'):
+            for project in self.get_jira_info(username, password, url, ssl_verify, proxy_dict ,'project'):
                 usercreds = {'_time': time.time(), 'key':project.get('key'), 'key_projects':project.get('key')+" - "+project.get('name')}
                 yield usercreds
 
         if self.opt == 2:
-            for issue in self.get_jira_info(username, password, url, proxy_dict , 'issuetype'):
+            for issue in self.get_jira_info(username, password, url, ssl_verify, proxy_dict , 'issuetype'):
                 usercreds = {'_time': time.time(), 'issues':issue.get('name')}
                 yield usercreds
 
         if self.opt == 3:
-            for priority in self.get_jira_info(username, password, url, proxy_dict , 'priority'):
+            for priority in self.get_jira_info(username, password, url, ssl_verify, proxy_dict , 'priority'):
                 usercreds = {'_time': time.time(), 'priorities':priority.get('name')}
                 yield usercreds
 
         if self.opt == 4:
-            for status in self.get_jira_info(username, password, url, proxy_dict , 'status'):
+            for status in self.get_jira_info(username, password, url, ssl_verify, proxy_dict , 'status'):
                 result = {'_time': time.time(), 'status':status.get('name'), 'statusCategory':status.get('statusCategory').get('name')}
                 yield result