+++ docs/releasenotes.rst
@@ -4,7 +4,8 @@ Release notes
 Version 1.0.30
 ==============
 
-- Enhancement: proxy support for jirarest.py and jirafill.py #91 (Author: 8lex)
+- Enhancement: Issue #91 - proxy support for jirarest.py and jirafill.py #91 (Author: 8lex)
+- Enhancement: Issue #92 - provide an SSL certificate path option for internal PKI certificate validation, honour SSL certificate validation in custom commands
 
 Version 1.0.29
 ==============